Minor improvements to build script

+++ src/PoshGram.build.ps1
@@ -151,13 +151,13 @@ task Test {
     if (Test-Path -Path $script:UnitTestsPath) {
         Write-Host -NoNewLine "      Performing Pester Unit Tests"
         $invokePesterParams = @{
-            Path         = 'Tests\Unit'
-            Strict       = $true
-            PassThru     = $true
-            Verbose      = $false
-            EnableExit   = $false
-            CodeCoverage = "$ModuleName\*\*.ps1"
-            CodeCoverageOutputFile = "$codeCovPath\codecoverage.xml"
+            Path                         = 'Tests\Unit'
+            Strict                       = $true
+            PassThru                     = $true
+            Verbose                      = $false
+            EnableExit                   = $false
+            CodeCoverage                 = "$ModuleName\*\*.ps1"
+            CodeCoverageOutputFile       = "$codeCovPath\codecoverage.xml"
             CodeCoverageOutputFileFormat = 'JaCoCo'
         }
 
@@ -197,11 +197,11 @@ task Test {
     if (Test-Path -Path $script:InfraTestsPath) {
         Write-Host -NoNewLine "      Performing Pester Infrastructure Tests"
         $invokePesterParams = @{
-            Path         = '..\..\Tests\Infrastructure'
-            Strict       = $true
-            PassThru     = $true
-            Verbose      = $false
-            EnableExit   = $false
+            Path       = '..\..\Tests\Infrastructure'
+            Strict     = $true
+            PassThru   = $true
+            Verbose    = $false
+            EnableExit = $false
         }
         Write-Host $invokePesterParams.path
         # Publish Test Results as NUnitXml
@@ -314,10 +314,13 @@ task UpdateCBH -Before Build {
 
 # Synopsis: Builds the Module to the Artifacts folder
 task Build {
-    Write-Host -NoNewLine "      Performing Module Build"
+    Write-Host "      Performing Module Build"
+
+    Write-Host '        Copying Module Manifest to Artifacts...'
     Copy-Item -Path $script:ModuleManifestFile -Destination $script:ArtifactsPath -Recurse -ErrorAction Stop
     #Copy-Item -Path $script:ModuleSourcePath\bin -Destination $script:ArtifactsPath -Recurse -ErrorAction Stop
 
+    Write-Host '        Merging Public and Private functions to one module file'
     #$private = "$script:ModuleSourcePath\Private"
     $scriptContent = [System.Text.StringBuilder]::new()
     #$powerShellScripts = Get-ChildItem -Path $script:ModuleSourcePath -Filter '*.ps1' -Recurse
@@ -327,13 +330,21 @@ task Build {
         $null = $scriptContent.AppendLine('')
         $null = $scriptContent.AppendLine('')
     }
-
     $scriptContent.ToString() | Out-File -FilePath $script:BuildModuleRootFile -Encoding utf8 -Force
 
+    Write-Host '        Cleaning up artifacts location'
     #cleanup artifacts that are no longer required
-    Remove-Item "$($script:ArtifactsPath)\Public" -Recurse -Force -ErrorAction Stop
-    Remove-Item "$($script:ArtifactsPath)\Private" -Recurse -Force -ErrorAction Stop
-    #Remove-Item "$($script:ArtifactsPath)\docs" -Recurse -Force -ErrorAction Stop
+    if (Test-Path "$($script:ArtifactsPath)\Public") {
+        Remove-Item "$($script:ArtifactsPath)\Public" -Recurse -Force -ErrorAction Stop
+    }
+    if (Test-Path "$($script:ArtifactsPath)\Private") {
+        Remove-Item "$($script:ArtifactsPath)\Private" -Recurse -Force -ErrorAction Stop
+    }
+
+    Write-Host "        Overwriting docs output"
+    Move-Item "$($script:ArtifactsPath)\docs\*.md" -Destination "..\docs\" -Force
+    Remove-Item "$($script:ArtifactsPath)\docs" -Recurse -Force -ErrorAction Stop
+
     Write-Host -ForegroundColor Green '...Build Complete!'
 }#Build
